requirements.txt
db-sqlite3==0.0.1
streamlit==1.31.1
ipython==8.15.0
bs4==0.0.1
google-generativeai==0.3.2
streamlit-chat==0.1.1
pandas==2.2.1
